Before the last commit
Continued working on index.html almost done with it. Need to fix the footer and make it responsive. Also add the portfolio pictures (just need to cut and paste not much work)

Went over some details with Taki to be changed.

Please do test it on your machine and report to me any incompatibility issues.